CHAPITRE-5/5.1-A-RENDRE/requirements.txt
Flask==3.0.3
Flask-Minify==0.48
Flask-WTF==1.2.1
mysql-connector-python==9.0.0
pillow==10.4.0
Pygments==2.18.0
python-dotenv==1.0.1
email-validator==2.2.0 
markdown

